fixed search form
general clean-up
