+++ index.html
@@ -7,7 +7,7 @@
   <title>MTG Collection</title>
 </head>
 
-<body>
+<body style="background-color: #d9d9d9">
   <div id="root"></div>
   <script type="module" src="/src/main.tsx"></script>
 </body>
