+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import Layout from './components/Layout';
-import RequireAuth from './components/RequireAuth';
-import Missing from './components/Missing';
+import Footer from './components/Footer';
+import RequireAuth from './pages/RequireAuth';
+import Missing from './pages/Missing';
 import Collection from './components/Collection';
 import Search from './components/Search';
 import Login from './components/Login';
@@ -18,17 +18,18 @@ function App() {
           <Routes>
             {/* Public */}
             <Route path="/login" element={<Login />} />
+            <Route path="/account" element={<Account />} />
 
             {/* Private */}
             <Route element={<RequireAuth />}>
               <Route path="/" element={<Collection />} />
               <Route path="/search" element={<Search />} />
-              <Route path="/account" element={<Account />} />
             </Route>
 
             {/* Catch All*/}
             <Route path="*" element={<Missing />} />
           </Routes>
+          <Footer />
           <ReactQueryDevtools initialIsOpen={false} />
       </BrowserRouter>
     </QueryClientProvider>
