--- src/components/LogoutMsg.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function LogoutMsg() {
-  return (
-    <div>LogoutMsg</div>
-  )
-}
-
-export default LogoutMsg
