+++ src/pages/RequireAuth.tsx
@@ -7,7 +7,7 @@ const RequireAuth = () => {
 
     auth?.accessToken ?
     <Outlet /> :
-    <Navigate to="/login" replace/>
+    <Navigate to="/login" replace={true}/>
   )
 }
